+++ packages/react-components/react-menu/src/contexts/menuTriggerContext.ts
@@ -3,8 +3,10 @@ import * as React from 'react';
 /**
  * Context used communicate with a child menu item that it is a trigger for a submenu
  */
-// eslint-disable-next-line @fluentui/no-context-default-value
-const MenuTriggerContext = React.createContext<boolean>(false);
+const MenuTriggerContext = React.createContext<boolean | undefined>(undefined) as React.Context<boolean>;
+
+const menuTriggerContextDefaultValue = false;
 
 export const MenuTriggerContextProvider = MenuTriggerContext.Provider;
-export const useMenuTriggerContext_unstable = () => React.useContext(MenuTriggerContext);
+export const useMenuTriggerContext_unstable = () =>
+  React.useContext(MenuTriggerContext) ?? menuTriggerContextDefaultValue;